requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SocketIO
numpy
pandas
scipy
python-dotenv
tiktoken
openai==1.3.4
langdetect==1.0.9
lingua-language-detector==2.0.2
PyPDF2
